+++ .eslintrc.js
@@ -19,7 +19,7 @@ module.exports = {
     'class-methods-use-this': 'off',
     'no-param-reassign': 'off',
     'no-undef': 'off',
-    'no-plusplus': 'off',
+
     camelcase: 'off',
     'no-unused-vars': ['error', { argsIgnorePattern: 'next' }],
     'linebreak-style': ['error', 'unix'],
