+++ src/services/movie/list.service.js
@@ -1,33 +1,13 @@
 const { Op } = require('sequelize');
 const { movieRepository } = require('../../repositories');
+const { pagination } = require('../../utils');
 
 module.exports.list = async (title, page, limit) => {
-  const result = [];
   const query = title !== '' ? { title: { [Op.iLike]: `%${title}%` } } : '';
 
   const movies = await movieRepository.find(query);
-  const totalPage = Math.ceil(movies.length / limit);
-  let count = page * limit - limit;
-  const delimiter = count + limit;
 
-  if (page <= totalPage) {
-    for (let i = count; i < delimiter; i++) {
-      if (movies[i] !== undefined) {
-        result.push(movies[i]);
-      }
-      count++;
-    }
-  }
-
-  const response = {
-    movies: result,
-    pagination: {
-      movies: movies.length,
-      limit,
-      totalPages: totalPage,
-      currentPage: page,
-    },
-  };
+  const response = pagination(page, limit, movies, 'movies');
 
   return response;
 };
